pyproject.toml
[project]
name = "orca-score"
version = "0.1.0"
description = "ORCA: Open-ended Response Correctness Assessment for Audio QAs"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Bolaji Yusuf"},
    {name = "Santosh Kesiraju"},
]
dependencies = [
    "torch",
    "transformers",
    "accelerate",
    "peft",
    "scipy",
    "pyyaml",
    "tqdm",
    "tensorboard",
    "bitsandbytes",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "isort",
    "ipdb",
]

[project.scripts]
orca-train = "orca_score.cli:orca_train"
orca-infer = "orca_score.cli:orca_infer"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orca_score"]

[tool.uv]
dev-dependencies = []
